Refactor: community layout 폭 맞게 다시 구현
