+++ src/pages/community/FriendPage.tsx
@@ -1,16 +1,16 @@
 import Friends from '@/components/community/friend/FriendList';
 import SearchFriends from '@/components/community/friend/SearchFriends';
 // import FriendModal from '@/modal/FriendModal';
-
+import CommunityPageLayout from '@/components/community/CommunityPageLayout';
 const FriendPage = () => {
   return (
-    <main className="w-full h-[calc(100vh-3.1875rem)] flex bg-Light_Layout-200 dark:bg-Dark_Layout-300 justify-center px-5 py-20 gap-8">
+    <CommunityPageLayout>
       <div className="flex gap-8">
         {/* <FriendModal type="block" /> */}
         <SearchFriends />
         <Friends />
       </div>
-    </main>
+    </CommunityPageLayout>
   );
 };
